quora-rag/requirements.txt
python-dotenv>=1.0.1
streamlit>=1.36.0
langchain>=0.2.10
langchain-openai>=0.1.17
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2
faiss-cpu>=1.8.0
openai>=1.40.0
tiktoken>=0.7.0
